TGU: 2FA for admins, 2-person mech, hair gradients. Fish price halved!! (#220)
* So much win!

* um

maptick_??????

+++ code/game/objects/structures/traps.dm
@@ -113,10 +113,12 @@
 	. = ..()
 	time_between_triggers = 10
 	flare_message = "<span class='warning'>[src] snaps shut!</span>"
-	var/static/list/loc_connections = list(
-		COMSIG_ATOM_ENTERED = .proc/on_entered,
-	)
-	AddElement(/datum/element/connect_loc, src, loc_connections)
+
+/obj/structure/trap/stun/hunter/Destroy()
+	if(!QDELETED(stored_item))
+		qdel(stored_item)
+	stored_item = null
+	return ..()
 
 /obj/structure/trap/stun/hunter/on_entered(datum/source, atom/movable/AM)
 	if(isliving(AM))
@@ -171,7 +173,9 @@
 	forceMove(stored_trap)//moves item into trap
 
 /obj/item/bountytrap/Destroy()
-	qdel(stored_trap)
+	if(!QDELETED(stored_trap))
+		qdel(stored_trap)
+	stored_trap = null
 	QDEL_NULL(radio)
 	QDEL_NULL(spark_system)
 	. = ..()
